add plugin extension and schema handle
