+++ src/plugins/github.ts
@@ -1,12 +1,16 @@
 import fs from 'fs-extra';
 import axios from 'axios';
-import type {DoculaPlugin} from '../docula-plugin.js';
+import type {DoculaPlugin, Options, Rules} from '../docula-plugin.js';
 import type {Config} from '../config.js';
-import type {GithubConfig} from '../types/config.js';
 import {type Runtime} from '../docula-plugin.js';
 
+export type GithubConfig = {
+	repo: string;
+	author: string;
+};
+
 export class GithubPlugin implements DoculaPlugin {
-	readonly options: Record<string, string> = {
+	readonly options: Options = {
 		api: 'https://api.github.com',
 		path: '_data',
 		author: '',
@@ -14,6 +18,14 @@ export class GithubPlugin implements DoculaPlugin {
 		outputFile: 'github.json',
 	};
 
+	static rules: Rules = {
+		type: 'object',
+		required: ['repo', 'author'],
+		properties: {
+			repo: {type: 'string'},
+			author: {type: 'string'},
+		},
+	}
 	runtime: Runtime = 'after';
 
 	constructor(config: Config) {
