+++ src/plugins/index.ts
@@ -1,8 +1,13 @@
-import {GithubPlugin} from './github.js';
-import {NpmPlugin} from './npm.js';
+import {type GithubConfig, GithubPlugin} from './github.js';
+import {type NpmConfig, NpmPlugin} from './npm.js';
 
 const plugins = {
 	github: GithubPlugin,
 	npm: NpmPlugin,
 };
+
+export type PluginConfig = GithubConfig | NpmConfig;
+
+export type PluginInstance = GithubPlugin | NpmPlugin;
+
 export default plugins;
