+++ src/plugins/npm.ts
@@ -1,15 +1,26 @@
 import axios from 'axios';
 import fs from 'fs-extra';
-import type {DoculaPlugin, Runtime} from '../docula-plugin.js';
+import type {DoculaPlugin, Options, Rules, Runtime} from '../docula-plugin.js';
 import type {Config} from '../config.js';
-import type {NpmConfig} from '../types/config.js';
 
-export class NpmPlugin implements DoculaPlugin {
-	readonly options: Record<string, string> = {
+export type NpmConfig = {
+	moduleName: string;
+}
+
+export class NpmPlugin implements DoculaPlugin  {
+	readonly options: Options = {
 		dataPath: '_data',
 		moduleName: '',
 		outputFile: 'npm.json',
 	};
+
+	static rules: Rules = {
+		type: 'object',
+		required: ['moduleName'],
+		properties: {
+			moduleName: {type: 'string'},
+		},
+	}
 
 	runtime: Runtime = 'before';
